+++ Jayme/Readable.swift
@@ -28,17 +28,29 @@ public protocol Readable: Repository {
 
 public extension Readable {
     
-    /// Returns a `Future` containing an array of all the entities in the repository, or the relevant `JaymeError` that could occur.
-    public func findAll() -> Future<[EntityType], JaymeError> {
+    /// Fetches the only entity from this repository
+    /// Returns a `Future` containing the only entity in the repository, or the relevant `JaymeError` that could occur.
+    /// Watch out for a `.failure` case with `JaymeError.entityNotFound`.
+    public func read() -> Future<EntityType, JaymeError> {
+        let path = self.name
+        return self.backend.future(path: path, method: .GET, parameters: nil)
+            .andThen { DataParser().dictionary(from: $0.0) }
+            .andThen { EntityParser().entity(from: $0) }
+    }
+    
+    /// Fetches all the entities from this repository
+    /// Returns a `Future` containing an array with all the entities in the repository, or the relevant `JaymeError` that could occur.
+    public func readAll() -> Future<[EntityType], JaymeError> {
         let path = self.name
         return self.backend.future(path: path, method: .GET, parameters: nil)
             .andThen { DataParser().dictionaries(from: $0.0) }
             .andThen { EntityParser().entities(from: $0) }
     }
     
+    /// Fetches only the entity from this repository that matches the given `id`
     /// Returns a `Future` containing the entity matching the `id`, or the relevant `JaymeError` that could occur.
     /// Watch out for a `.failure` case with `JaymeError.entityNotFound`.
-    public func find(byId id: EntityType.IdentifierType) -> Future<EntityType, JaymeError> {
+    public func read(id: EntityType.IdentifierType) -> Future<EntityType, JaymeError> {
         let path = "\(self.name)/\(id)"
         return self.backend.future(path: path, method: .GET, parameters: nil)
             .andThen { DataParser().dictionary(from: $0.0) }