+++ JaymeTests/CreatableTests.swift
@@ -37,11 +37,13 @@ class CreatableTests: XCTestCase {
     
 }
 
+// MARK: - Create Single Entity
+
 extension CreatableTests {
     
     // MARK: - Call To backend
     
-    func testCreateCall() {
+    func testCreateEntityCall() {
         let document = TestDocument(id: "123", name: "a")
         let _ = self.repository.create(document)
         XCTAssertEqual(self.backend.path, "documents")
@@ -58,7 +60,7 @@ extension CreatableTests {
     
     // MARK: - Success Response
     
-    func testCreateSuccessCallback() {
+    func testCreateEntitySuccessCallback() {
         
         // Simulated completion
         self.backend.completion = { completion in
@@ -83,10 +85,10 @@ extension CreatableTests {
             if let _ = error { XCTFail() }
         }
     }
-
+    
     // MARK: - Failure Response
     
-    func testCreateFailureCallback() {
+    func testCreateEntityFailureCallback() {
         
         // Simulated completion
         self.backend.completion = { completion in
@@ -109,4 +111,87 @@ extension CreatableTests {
         }
     }
 
+}
+
+// MARK: - Create Multiple Entities
+
+extension CreatableTests {
+    
+    // MARK: - Call To backend
+    
+    func testCreateEntitiesCall() {
+        let doc1 = TestDocument(id: "1", name: "doc1")
+        let doc2 = TestDocument(id: "2", name: "doc2")
+        let _ = self.repository.create([doc1, doc2])
+        XCTAssertEqual(self.backend.path, "documents")
+        XCTAssertEqual(self.backend.method, .POST)
+        XCTAssertEqual(self.backend.parametersAsArray?.count, 2)
+        guard
+            let id1 = self.backend.parametersAsArray?[0]["id"] as? String,
+            let name1 = self.backend.parametersAsArray?[0]["name"] as? String,
+            let id2 = self.backend.parametersAsArray?[1]["id"] as? String,
+            let name2 = self.backend.parametersAsArray?[1]["name"] as? String
+            else { XCTFail("Wrong parameters"); return }
+        XCTAssertEqual(id1, "1")
+        XCTAssertEqual(name1, "doc1")
+        XCTAssertEqual(id2, "2")
+        XCTAssertEqual(name2, "doc2")
+    }
+    
+    // MARK: - Success Response
+    
+    func testCreateEntitiesSuccessCallback() {
+        
+        // Simulated completion
+        self.backend.completion = { completion in
+            let json = [["id": "1", "name": "a"], ["id": "2", "name": "b"]]
+            let data = try! JSONSerialization.data(withJSONObject: json, options: .prettyPrinted)
+            completion(.success((data, nil)))
+        }
+        
+        let expectation = self.expectation(description: "Expected to get a success")
+        
+        let documents: [TestDocument] = []
+        let future = self.repository.create(documents)
+        future.start() { result in
+            guard case .success(let docs) = result
+                else { XCTFail(); return }
+            XCTAssertEqual(docs.count, 2)
+            XCTAssertEqual(docs[0].id, "1")
+            XCTAssertEqual(docs[0].name, "a")
+            XCTAssertEqual(docs[1].id, "2")
+            XCTAssertEqual(docs[1].name, "b")
+            expectation.fulfill()
+        }
+        
+        self.waitForExpectations(timeout: 3) { error in
+            if let _ = error { XCTFail() }
+        }
+    }
+    
+    // MARK: - Failure Response
+    
+    func testCreateEntitiesFailureCallback() {
+        
+        // Simulated completion
+        self.backend.completion = { completion in
+            let error = JaymeError.notFound
+            completion(.failure(error))
+        }
+        
+        let expectation = self.expectation(description: "Expected to get an error")
+        
+        let documents: [TestDocument] = []
+        let future = self.repository.create(documents)
+        future.start() { result in
+            guard case .failure = result
+                else { XCTFail(); return }
+            expectation.fulfill()
+        }
+        
+        self.waitForExpectations(timeout: 3) { error in
+            if let _ = error { XCTFail() }
+        }
+    }
+    
 }